requirements.txt
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==11.3.0
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.30.0
kornia>=0.7.0
timm>=0.9.0
fastapi==0.112.2
python-multipart==0.0.20
uvicorn==0.34.3
pydantic-settings>=2.0.0